+++ app/src/renderer/components/common/ActionModal.vue
@@ -130,6 +130,7 @@ import TmFormGroup from "common/TmFormGroup"
 import TmFormMsg from "common/TmFormMsg"
 import { mapGetters } from "vuex"
 import { requiredIf } from "vuelidate/lib/validators"
+import { track } from "../../google-analytics.js"
 
 const defaultStep = `txDetails`
 const signStep = `sign`
@@ -202,6 +203,8 @@ export default {
   },
   methods: {
     open() {
+      track(`event`, `modal`, this.title)
+
       this.show = true
     },
     close() {
@@ -239,6 +242,8 @@ export default {
       }
     },
     async submit() {
+      track(`event`, `submit`, this.title, this.selectedSignMethod)
+
       try {
         await this.submitFn(this.selectedSignMethod, this.password)