+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- [\#1032](https://github.com/cosmos/voyager/issues/1032) added several events to google analytics @faboweb
+
 ## [1.0.0-beta.7] - 2019-02-26
 
 ### Changed